refactor: move mesh agents to `agents` folder
